@@ -5,20 +5,20 @@
     "keywords": ["framework", "laravel"],
     "license": "MIT",
     "require": {
-        "php": "^7.3|^8.0",
-        "fruitcake/laravel-cors": "^2.0",
-        "guzzlehttp/guzzle": "^7.0.1",
-        "laravel/framework": "^8.75",
-        "laravel/sanctum": "^2.11",
-        "laravel/tinker": "^2.5"
+        "php": "^8.2",
+        "guzzlehttp/guzzle": "^7.7",
+        "laravel/framework": "^10.0",
+        "laravel/sanctum": "^3.2",
+        "laravel/tinker": "^2.8"
     },
     "require-dev": {
-        "facade/ignition": "^2.5",
-        "fakerphp/faker": "^1.9.1",
-        "laravel/sail": "^1.0.1",
-        "mockery/mockery": "^1.4.4",
-        "nunomaduro/collision": "^5.10",
-        "phpunit/phpunit": "^9.5.10"
+        "spatie/laravel-ignition": "^2.0",
+        "fakerphp/faker": "^1.21",
+        "laravel/breeze": "^1.10",
+        "laravel/sail": "^1.22",
+        "mockery/mockery": "^1.5",
+        "nunomaduro/collision": "^6.4",
+        "phpunit/phpunit": "^10.2"
     },
     "autoload": {
         "psr-4": {
